google-cloud-brain/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.14.0
google-cloud-vision==3.5.0
google-cloud-translate==3.12.1
google-cloud-speech==2.21.0
python-multipart==0.0.6
